refactor(cron): split service timer helpers
